--- a/Misc/NEWS.d/next/Library/2025-02-09-16-21-42.gh-issue-118761.ryMtuz.rst
+++ b/Misc/NEWS.d/next/Library/2025-02-09-16-21-42.gh-issue-118761.ryMtuz.rst
@@ -1,0 +1,2 @@
+Improve import time of :mod:`dataclasses` by lazy importing :mod:`re`,
+:mod:`copy`, :mod:`inspect` and :mod:`annotationlib`. Patch by Semyon Moroz.