@donBarbos 2025-02-10
## On Trade-offs

### Here is the new call trace

1. `import re;re.compile` in `_is_type()`
2. `_is_type()` in `_process_class()` and `_get_field()`
3. `_get_field()` in `_process_class()`
4. `_process_class()` in `dataclass.wrap()`

---

1. `import annotationlib` in `_process_class()`
2. `_process_class()` in `dataclass.wrap()`

---

1. `import inspect` in `_process_class()` and `_add_slots()`
2. `_add_slots()` in `_process_class()`
3. `_process_class()` in `dataclass.wrap()`

---

1. `import copy` in `_asdict_inner()` and `_astuple_inner()`
2. `_asdict_inner()` in `asdict()`
3. `_astuple_inner()` in `astuple()`

---

And I wrote benchmarks to test calling affected public functions (`dataclass` decorator and `asdict`, `astuple` functions):
### `bench_dataclass.py` starts like this:
```python
import time
import dataclasses
import os

code = """
@dataclasses.dataclass
class Address:
    city: str
    zip_code: str

@dataclasses.dataclass
class Person:
    name: str
    age: int
    addresses: list[Address] = dataclasses.field(default_factory=list)
    metadata: dict[str, str] = dataclasses.field(default_factory=dict)
"""

result_times = []

for _ in range(100):
    os.system("sync && echo 3 | sudo tee /proc/sys/vm/drop_caches > /dev/null")
    start_time = time.time()
    exec(code, {"dataclasses": dataclasses})
    end_time = time.time()
    result_times.append(end_time-start_time)
```

### `bench_asdict.py` starts like this:
```python
import time
import dataclasses
import os

@dataclasses.dataclass
class Address:
    city: str
    zip_code: str

@dataclasses.dataclass
class Person:
    name: str
    age: int
    addresses: list[Address] = dataclasses.field(default_factory=list)
    metadata: dict[str, str] = dataclasses.field(default_factory=dict)

person = Person(
    name="John Doe",
    age=30,
    addresses=[Address("New York", "10001"), Address("Los Angeles", "90001")],
    metadata={"key1": "value1", "key2": "value2"}
)

code = """
dataclasses.asdict(person)
"""

result_times = []

for _ in range(100):
    os.system("sync && echo 3 | sudo tee /proc/sys/vm/drop_caches > /dev/null")
    start_time = time.time()
    exec(code, {"dataclasses": dataclasses, "person": person})
    end_time = time.time()
    result_times.append(end_time-start_time)
```

### And I getting stats in both scripts in the end of file like this:
```python
import statistics

first = result_times[0]
mean = statistics.mean(result_times)
median = statistics.median(result_times)
stdev = statistics.stdev(result_times)
variance = statistics.variance(result_times)

print(f"First time: {first * 1_000_000:.2f}μs")
print(f"Mean: {mean * 1_000_000:.2f}μs")
print(f"Median: {median * 1_000_000:.2f}μs")
print(f"Standard deviation: {stdev * 1_000_000:.2f}μs")
print(f"Variance: {variance * 1_000_000:.2f}μs")
```

### Results on main branch:
```shell
$ ./python -B bench_dataclass.py
First time: 1586.20μs
Mean: 1267.27μs
Median: 1243.83μs
Standard deviation: 106.15μs
Variance: 0.01μs
$ ./python -B bench_asdict.py
First time: 99.66μs
Mean: 112.33μs
Median: 109.91μs
Standard deviation: 12.26μs
Variance: 0.00μs
```
### Results on PR branch:
```shell
$ ./python -B bench_dataclass.py
First time: 20320.18μs  # is 20ms or 0.02s and this is a permanent result for first call
Mean: 1454.47μs
Median: 1243.11μs
Standard deviation: 1911.43μs
Variance: 3.65μs
$ ./python -B bench_asdict.py
First time: 100.37μs
Mean: 112.21μs
Median: 108.96μs
Standard deviation: 13.01μs
Variance: 0.00μs
```

In the end we can say that about 20ms was added for the first mention of `@dataclass` in the code.