@eli-schwartz 2025-02-16
I'm not sure what point you're trying to make.

Most of the PRs for this issue are adding lazy imports. Lazy imports are a useful tool for making python programs faster, in the principle of "only pay for what you use" -- and stdlib modules often don't know what an application will in fact use. 

The issue here is that as far as I can tell you're adding lazy imports for things that the consumer will *always* use, which means that there won't be a benefit to making them lazy...